+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## [0.0.2] - Add more options to all widgets 
+## [0.0.2+] - Fix future builder
+
+## [0.0.2] - Add more options to all widgets
 
 ## [0.0.1+1] - Add Future builder.